# src/appconfigs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Welcome Agri-Vesty API",
    "description": "Agri-vesty is a crowdfunding Platform. The aim of Agri-vesty is to help farmers to raise farming capital. In this platform, a user can either create a crowdfunding farming project or invest money towards their favorite farming projects.",
    "contact": {
      "name": "Team-179-Group-A",
      "url": "https://agri-vesty.herokuapp.com/"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:{port}/{basePath}",
      "description": "Local server",
      "variables": {
        "port": {
          "enum": [
            "8443",
            "443"
          ],
          "default": "4000"
        },
        "basePath": {
          "default": "api"
        }
      }
    },
    {
      "url": "https://agri-vesty.herokuapp.com/{basePath}",
      "description": "Production server",
      "variables": {
        "basePath": {
          "default": "api"
        }
      }
    }
  ],
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "components": {
    "schemas": {
...
      "Profile": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string"
          },
          "bios": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "idNumber": {
            "type": "string"
          },
          "kraPin": {
            "type": "string"
          },
          "certificateOfConduct": {
            "type": "string"
          }
        }
      },
...
    },
...
    "responses": {
      "Error": {
        "description": "General Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GeneralError"
            }
          }
        }
      },
...
    },
    "securitySchemes":{
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "Bearer": {
        "description": "JWT Token",
        "name": "Authorization",
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/user/create-profile": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "createProfile",
        "summary": "Create user additional information",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "bios": {
                    "type": "string"
                  },
                  "roles": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "idNumber": {
                    "type": "string"
                  },
                  "kraPin": {
                    "type": "string"
                  }
                },
                "required": [
                  "phoneNumber",
                  "gender",
                  "dateOfBirth",
                  "idNumber",
                  "kraPin"
                ]
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "images": {
                    "type": "string"
                  },
                  "bios": {
                    "type": "string"
                  },
                  "roles": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date",
                    "example": "2018-06-22"
                  },
                  "idNumber": {
                    "type": "string"
                  },
                  "kraPin": {
                    "type": "string"
                  }
                },
                "required": [
                  "phoneNumber",
                  "gender",
                  "dateOfBirth",
                  "idNumber",
                  "kraPin"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New profile created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "success": true,
                  "message": "Profile created succeessfully",
                  "data": {
                    "$ref": "#/components/schemas/Profile"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                },
                "example": {
                  "message": "No user to associate with that profile",
                  "internal_code": "missing_parameters"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/user/update-profile/{profileId}": {
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "updateProfile",
        "summary": "Update user information",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "description": "User Profile ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "uniqueItems": true
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "roles": {
                    "type": "string"
                  },
                  "bios": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "idNumber": {
                    "type": "string"
                  },
                  "kraPin": {
                    "type": "string"
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "roles",
                  "phoneNumber",
                  "gender",
                  "dateOfBirth",
                  "idNumber",
                  "kraPin"
                ]
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "uniqueItems": true
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "roles": {
                    "type": "string"
                  },
                  "bios": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "idNumber": {
                    "type": "string"
                  },
                  "kraPin": {
                    "type": "string"
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "roles",
                  "phoneNumber",
                  "gender",
                  "dateOfBirth",
                  "idNumber",
                  "kraPin"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "success": true,
                  "message": "Profile updated succeessfully"
                }
              }
            }
          },
          "404": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/user/change-password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "changePassword",
        "summary": "Change user password",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "format": "password"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "required": [
                  "oldPassword",
                  "password",
                  "confirmPassword"
                ]
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "format": "password"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "required": [
                  "password",
                  "confirmPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Changed Password successfully",
            "content": {
              "application/json": {
                "schema": {
                  "success": true,
                  "message": "Changed Password successfully"
                }
              }
            }
          },
          "404": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/user/forgot-password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "forgotPassword",
        "summary": "reset password",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "uniqueItems": true
                  }
                },
                "required": [
                  "email"
                ]
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "uniqueItems": true
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent to your email",
            "content": {
              "application/json": {
                "schema": {
                  "success": true,
                  "message": "Email sent to your email"
                }
              }
            }
          },
          "404": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/password-reset": {
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "resetPassword",
        "summary": "Reset user password",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "uniqueItems": true
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "format": "password"
                  },
                  "code": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "confirmPassword",
                  "code"
                ]
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "uniqueItems": true
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "format": "password"
                  },
                  "code": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "confirmPassword",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset Password successful",
            "content": {
              "application/json": {
                "schema": {
                  "success": true,
                  "message": "Reset Password successful"
                }
              }
            }
          },
          "404": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/delete": {
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "deleteUser",
        "summary": "Delete User",
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "success": true,
                  "message": "User deleted suscessfully"
                }
              }
            }
          },
          "403": {
            "description": "Forbiden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  }
}